# openapi/openapi.yaml
...
openapi: 3.0.3
info:
    contact:
        name: metarang
    description: REST API for metarang microservices. Legacy-compatible JSON responses served via Kong → service HTTP APIs.
    title: metarang API
    version: 1.0.0
servers:
    - description: Kong API Gateway (local)
      url: http://localhost:8000
    - description: support-service HTTP direct (dev)
      url: http://localhost:8070
...
paths:
...
    /api/notifications:
        get:
            operationId: get_notifications
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: List notifications
            tags:
                - Notifications
    /api/notifications/read/{id}:
        post:
            operationId: post_notifications_read_id
            parameters:
                - description: Notification ID
                  in: path
                  name: id
                  required: true
                  schema:
                    example: 550e8400-e29b-41d4-a716-446655440000
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                    multipart/form-data:
                        schema:
                            type: object
                required: true
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: Mark notification as read
            tags:
                - Notifications
    /api/notifications/read/all:
        post:
            operationId: post_notifications_read_all
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                    multipart/form-data:
                        schema:
                            type: object
                required: true
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: Mark all notifications as read
            tags:
                - Notifications
    /api/order:
        post:
            operationId: post_order
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                    multipart/form-data:
                        schema:
                            type: object
                required: true
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: Create payment order
            tags:
                - Financial
    /api/order/callback:
        get:
            operationId: get_order_callback
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            summary: Payment gateway callback
            tags:
                - Financial
        post:
            operationId: post_order_callback
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                    multipart/form-data:
                        schema:
                            type: object
                required: true
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            summary: Payment gateway callback
            tags:
                - Financial
...
    /api/store:
        get:
            operationId: get_store
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - {}
                - BearerAuth: []
            summary: List store packages
            tags:
                - Financial
...
components:
    schemas:
...
        Error:
            properties:
                message:
                    type: string
            type: object
...
        ValidationError:
            properties:
                errors:
                    additionalProperties:
                        items:
                            type: string
                        type: array
                    type: object
                message:
                    type: string
            type: object
    securitySchemes:
        BearerAuth:
            bearerFormat: Token
            description: 'Bearer token. Header: Authorization: Bearer {token}'
            scheme: bearer
            type: http